Extract deprecated strategies from Clearance
These password strategies are deprecated and will be removed from
Clearance in 2.0. We do not intend to continue maintaining them and do
not encourage their use, thus the `deprecated` in the gem name.

Clearance will give deprecation warnings when attempting to use these
strategies. Adding this gem to the Gemfile will allow the continued use
of these strategies without the deprecation.